Move Worker to utils, fix tests
