--- mslib/worker.py
@@ -1,44 +1,0 @@
-"""
-    mslib.worker
-    ~~~~~~~~~~~~~~~~~~~
-    This module contains the "Worker" class, which can be used to run
-    a function through a QThread without much struggle, and receive the
-    return value or exception through signals.
-    Beware not to modify the parents connections through the function.
-
-    This file is part of mss.
-
-    :copyright: Copyright 2021 May Bär
-    :copyright: Copyright 2021 by the mss team, see AUTHORS.
-    :license: APACHE-2.0, see LICENSE for details.
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-"""
-
-from PyQt5.QtCore import QThread, pyqtSignal
-
-
-class Worker(QThread):
-    finished = pyqtSignal(object)
-    failed = pyqtSignal(Exception)
-
-    def __init__(self, function):
-        super(Worker, self).__init__()
-        self.function = function
-
-    def run(self):
-        try:
-            result = self.function()
-            self.finished.emit(result)
-        except Exception as e:
-            self.failed.emit(e)
